patch route fully working and errors are accounted for
