+++ lib/decodeToken.ts
@@ -1,6 +1,17 @@
 import { NextRequest, NextResponse } from "next/server";
 import jwt from "jsonwebtoken";
 
+export type DecodedToken = {
+  id: string;
+  username: string;
+  email: string;
+  handle: string;
+  bio: string;
+  role: string;
+  password_digest: string;
+  status: string;
+};
+
 export const getDataFromToken = (request: NextRequest) => {
   try {
     const token = request.cookies.get("token")?.value || "";
@@ -10,5 +21,23 @@ export const getDataFromToken = (request: NextRequest) => {
     return decodedToken;
   } catch (error: any) {
     throw new Error(error.message);
+  }
+};
+
+export const decodeUserToken = async (request: NextRequest) => {
+  try {
+    const token = request.cookies.get("token")?.value || "";
+
+    if (!token) {
+      return null;
+    }
+
+    const decodedToken = jwt.verify(
+      token,
+      process.env.JWT_SECRET
+    ) as DecodedToken;
+    return decodedToken;
+  } catch (error) {
+    return null;
   }
 };
